requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
google-generativeai
xlsxwriter
requests
plotly
streamlit
scipy
pydantic
python-multipart